backend/app/models/chat.py
from pydantic import BaseModel
from typing import List, Optional

class ChatMessage(BaseModel):
    message: str

class ChatResponse(BaseModel):
    response: str
    sources: Optional[List[str]] = None

class HealthResponse(BaseModel):
    status: str
    message: str