backend/requirements.txt
# Web framework
fastapi==0.115.4
uvicorn[standard]==0.32.0

# LlamaIndex - Python 3.12 compatible versions
llama-index==0.10.68
llama-index-core==0.10.68
llama-index-embeddings-huggingface==0.2.3
llama-index-vector-stores-qdrant==0.2.14
llama-index-llms-gemini==0.1.11

# Qdrant client
qdrant-client==1.12.1

# HuggingFace & embeddings
transformers==4.46.2
sentence-transformers==3.2.1
torch==2.5.1

# Environment variables
python-dotenv==1.0.1

# Utils
requests==2.32.3

# For async support in FastAPI
httpx==0.27.2

# Local package installation
-e .